functions/requirements.txt
azure-functions==1.14.0
azure-storage-blob==12.19.1
azure-cosmos==4.6.0
requests==2.31.0
opencensus-ext-azure==1.0.7
opencensus==0.11.2
